task#37636 - перевод Queue/DTO/Product->quantity с int на float

+++ examples/product_example.php
@@ -136,7 +136,7 @@ $supplierInfoDtoItem = new SupplierInfo(
 $productDto = new Product(
     (string)$product['name'],
     $price,
-    (int)$product['quantity'],
+    (float)$product['quantity'],
     $priceSum,
     (string)$product['tax'],
     $taxSum,
